+++ src/app/components/LoginForm/LoginForm.module.css
@@ -1,4 +1,4 @@
-/* .login{
+.login{
     display: flex;
     flex-direction: column;    
     justify-content: center;
@@ -10,10 +10,13 @@
     display: flex;
     flex-direction: column;
     justify-content: space-between;
-    width: 30%;    
+    /* width: 30%;     */
+    width: 100%;
+    height: 77vh;
     gap: 20px;    
-    padding: 40px;
-    border-radius: 10px;
+    padding: 38px;
+    border-radius: 0 10px 10px 0;
+    /* background: #5A72A0;     */
     background: white;
 }
 
@@ -50,4 +53,4 @@
 .redirect{
     display: flex;
     align-items: center;
-} */
+}
