+++ src/app/components/NavBar/NavBar.module.css
@@ -7,6 +7,7 @@
     gap: 20px;
     position: fixed;        
     backdrop-filter: blur(10px);    
+    z-index: 100;
 }
 
 .title{
