+++ src/app/dashboard/page.module.css
@@ -1,3 +1,4 @@
 .childrenContainer{
     padding-top: 120px;
+    height: 100vh;    
 }
