--- src/app/dashboard/services/[id]/pageStyles.module.css
@@ -1,6 +1,0 @@
-.mainPage{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-}
